requirements.txt
aiofiles
aiohttp[speedups]
async_generator
asyncurban
beautifulsoup4
bitlyshortener
bs4
carbonnow
cffi
colour
cloudscraper
covid
cowpy
cryptg
deezloader
dnspython
emoji
faker
ffmpeg-python
fontTools
geopy
git+https://github.com/yt-dlp/yt-dlp@master
gitpython
glitch_this
google_images_download
google_trans_new
google-api-python-client
googletrans==4.0.0rc1
gtts
hachoir
hentai
heroku3
html_telegraph_poster
htmlwebshot
httplib2
humanize 
jikanpy
js2py
justwatch
lottie
lxml
lyricsgenius
markdown
motor
natsort
nekos.py
oauth2client
opencv-python
pendulum
pillow
psutil
psycopg2-binary
py-tgcalls==0.8.3
pyaztro
pybase64
pyfiglet
PyGithub
pylast
PyPDF2
pySmartDL
python-barcode
python-dateutil
python-dotenv
pytube>=9.5.1
pytz
qrcode
rarfile
randomstuff.py
regex
requests
scipy
search-engine-parser
selenium
setuptools
spamwatch
speedtest-cli
sqlalchemy-json
sqlalchemy==1.3.23
telegraph
telethon
telethon-session-sqlalchemy
telethon-tgcrypto
urbandict
validators
vcsi
wand
wget
wikipedia
wordcloud
youtube-dl
youtube-search
youtube-search-python==1.4.9
zipfile38
